requirements.txt
requests
python-dotenv
langchain_openai
langchain-core
langgraph
opencv-python
mss
sounddevice 
numpy
websockets
fastapi
google-genai>=0.8.0
pdfminer
python-multipart
pdfminer.six
pyaudio
pydub
uvicorn[standard]
pymongo
motor